cloud-paper-news/requirements.txt
Django==1.7.1
Fabric==1.10.0
Pillow==2.6.1
argparse==1.2.1
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.4.6
django-storages==1.1.8
django-toolbelt==0.0.1
ecdsa==0.11
gunicorn==19.1.1
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
redis==2.10.3
six==1.8.0
sorl-thumbnail==11.12.1b
sqlparse==0.1.13
static3==0.5.1
wsgiref==0.1.2
